+++ main.go
@@ -7,6 +7,7 @@ package main
 
 import (
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"runtime"
@@ -22,7 +23,8 @@ import (
 )
 
 func main() {
-	var isCLI bool
+
+	var fyneApp fyne.App
 	// handle application start: CLI, GUI
 	args := len(os.Args)
 	if args > 1 && os.Args[1] == "cli" {
@@ -30,51 +32,103 @@ func main() {
 			fmt.Println("CLI app is unsupported on this OS")
 			os.Exit(1)
 		}
-		isCLI = true
+	} else {
+		fyneApp = app.NewWithID(ui.AppID)
+		fyneApp.SetIcon(icon.PawIcon)
+		if runtime.GOOS == "windows" {
+			// On Windows, to ship a single binary for GUI and CLI we need to build as
+			// "console binary" and detach the console when running as GUI
+			ui.DetachConsole()
+		}
 	}
 
-	if !isCLI && runtime.GOOS == "windows" {
-		// On Windows, to ship a single binary for GUI and CLI we need to build as
-		// "console binary" and detach the console when running as GUI
-		ui.DetachConsole()
-	}
-
-	s, err := paw.NewOSStorage()
+	s, err := makeStorage(fyneApp)
 	if err != nil {
-		log.Fatal(err)
+		fmt.Println(err)
+		os.Exit(1)
+	}
+
+	// Enable debugging log, if set
+	wc := initDebugLog(s)
+	defer wc.Close()
+
+	if fyneApp == nil {
+		// make and run the CLI app
+		cmd, err := cli.New(s)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "[✗] %s\n", err)
+			os.Exit(1)
+		}
+		err = cmd.Run(s)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "[✗] %s\n", err)
+			os.Exit(1)
+		}
+		return
 	}
 
 	// check for running instance
+	if ui.HealthServiceCheck(s.LockFilePath()) {
+		msg := "Paw GUI is already running\n"
+		wc.Write([]byte(msg))
+		os.Exit(1)
+	}
+	go ui.HealthService(s.LockFilePath())
+
+	// agent could be already running (e.g. from CLI)
+	// if not, start it
 	var agentType agent.Type
 	c, err := agent.NewClient(s.SocketAgentPath())
 	if err == nil {
 		agentType, _ = c.Type()
 	}
 
-	// handle application start: CLI, GUI
-	if isCLI {
-		// make CLI app
-		cli.New(s)
-		return
-	}
-
-	if ui.HealthServiceCheck(s.LockFilePath()) {
-		fmt.Println("Paw GUI is already running")
-		os.Exit(1)
-	}
-
-	go ui.HealthService(s.LockFilePath())
-
 	if agentType.IsZero() {
 		go agent.Run(agent.NewGUI(), s.SocketAgentPath())
 	}
 
-	a := app.NewWithID("dev.lucor.paw")
-	a.SetIcon(icon.PawIcon)
-
-	w := a.NewWindow("Paw")
+	// create window and run the app
+	w := fyneApp.NewWindow(ui.AppTitle)
 	w.SetMaster()
 	w.Resize(fyne.NewSize(400, 600))
 	w.SetContent(ui.MakeApp(w))
 	w.ShowAndRun()
+}
+
+// makeStorage create the storage
+func makeStorage(fyneApp fyne.App) (paw.Storage, error) {
+	if fyneApp == nil {
+		// CLI app returns the OS storage
+		return paw.NewOSStorage()
+	}
+	device := fyneApp.Driver().Device()
+	if device.IsMobile() {
+		// Fyne Mobile app returns the Fyne storage
+		return paw.NewFyneStorage(fyneApp.Storage())
+	}
+	// Fyne Desktop app returns the OS storage
+	return paw.NewOSStorage()
+}
+
+// initDebugLog create the debug log, if enabled
+func initDebugLog(s paw.Storage) io.WriteCloser {
+	w := os.Stderr
+	// Enable log debugging
+	if !paw.IsDebug() {
+		return w
+	}
+	// init logger
+	if fyne.CurrentDevice().IsMobile() {
+		w := os.Stdout
+		log.SetOutput(w)
+		log.SetPrefix("[paw] ")
+		return w
+	}
+	w, err := os.OpenFile(s.LogFilePath(), os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0644)
+	if err != nil {
+		fmt.Fprintln(w, "paw: error writing log file %w", err)
+		return w
+	}
+	log.SetOutput(w)
+	return w
 }